refact: method 위치 변경
