+++ src/repository/channel-member.repository.ts
@@ -5,29 +5,4 @@ import { IFindChannelMemberResult } from 'src/structure/dto/Channel';
 @Injectable()
 export class ChannelMemberRepository {
   constructor(private readonly prisma: PrismaService) {}
-
-  public async findChannelMembers(channelId: string) {
-    return this.prisma.gh_MemberInChannel
-      .findMany({
-        where: { channel_id: BigInt(channelId) },
-        include: {
-          user: {
-            select: {
-              id: true,
-              nickname: true,
-              avatar_url: true,
-            },
-          },
-        },
-      })
-      .then((rst) =>
-        rst.map(({ user }) => {
-          return {
-            id: user.id.toString(),
-            nickname: user.nickname,
-            avatar_url: user.avatar_url || '',
-          } satisfies IFindChannelMemberResult;
-        }),
-      );
-  }
 }
